requirements.txt
streamlit
firebase-admin
google-cloud-storage
python-dotenv
Pillow
pydub
qrcode
numpy
